// deno/gRPC/area_calculator/area_calculator.d.ts
/* this code was generated by automated tool, 
   should not edit by hand */

export interface Calculator {
  calculateOne(request: ShapeMessage): Promise<AreaResponse>;
  calculateMulti(request: AreaRequest): Promise<AreaResponse>;
}

export interface ShapeMessage {
  square?: Square;
  rectangle?: Rectangle;
  circle?: Circle;
  triangle?: Triangle;
  parallelogram?: Parallelogram;
}

export interface Square {
  edgeLength?: number;
}

export interface Rectangle {
  length?: number;
  width?: number;
}

export interface Circle {
  radius?: number;
}

export interface Triangle {
  edgeA?: number;
  edgeB?: number;
  edgeC?: number;
}

export interface Parallelogram {
  baseLength?: number;
  height?: number;
}

export interface AreaRequest {
  shapes?: ShapeMessage[];
}

export interface AreaResponse {
  value?: number[];
}